# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/transcription/{transcription_id}/minutes": {
      "get": {
        "tags": [
          "Minutes"
        ],
        "summary": "List Minutes For Transcription",
        "operationId": "list_minutes_for_transcription_transcription__transcription_id__minutes_get",
        "parameters": [
          {
            "name": "transcription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transcription Id"
            }
          },
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MinuteListItem"
                  },
                  "title": "Response List Minutes For Transcription Transcription  Transcription Id  Minutes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Minutes"
        ],
        "summary": "Create Minute",
        "operationId": "create_minute_transcription__transcription_id__minutes_post",
        "parameters": [
          {
            "name": "transcription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transcription Id"
            }
          },
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MinutesCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinuteVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/minutes/{minutes_id}": {
      "get": {
        "tags": [
          "Minutes"
        ],
        "summary": "Get Minute",
        "operationId": "get_minute_minutes__minutes_id__get",
        "parameters": [
          {
            "name": "minutes_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Minutes Id"
            }
          },
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Minute"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/minutes/{minute_id}/versions": {
      "get": {
        "tags": [
          "Minutes"
        ],
        "summary": "List Minute Versions",
        "operationId": "list_minute_versions_minutes__minute_id__versions_get",
        "parameters": [
          {
            "name": "minute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Minute Id"
            }
          },
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MinuteVersionResponse"
                  },
                  "title": "Response List Minute Versions Minutes  Minute Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Minutes"
        ],
        "summary": "Create Minute Version",
        "operationId": "create_minute_version_minutes__minute_id__versions_post",
        "parameters": [
          {
            "name": "minute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Minute Id"
            }
          },
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MinuteVersionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinuteVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/minute_versions/{minute_version_id}": {
      "get": {
        "tags": [
          "Minutes"
        ],
        "summary": "Get Minute Version",
        "operationId": "get_minute_version_minute_versions__minute_version_id__get",
        "parameters": [
          {
            "name": "minute_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Minute Version Id"
            }
          },
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinuteVersion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Minutes"
        ],
        "summary": "Delete Minute Version",
        "operationId": "delete_minute_version_minute_versions__minute_version_id__delete",
        "parameters": [
          {
            "name": "minute_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Minute Version Id"
            }
          },
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/minute_versions/{minute_version_id}/guardrails": {
      "get": {
        "tags": [
          "Minutes"
        ],
        "summary": "Get Guardrail Warning",
        "operationId": "get_guardrail_warning_minute_versions__minute_version_id__guardrails_get",
        "parameters": [
          {
            "name": "minute_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Minute Version Id"
            }
          },
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardrailWarningResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AiEdit": {
        "properties": {
          "instruction": {
            "type": "string",
            "title": "Instruction"
          },
          "source_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Id"
          }
        },
        "type": "object",
        "required": [
          "instruction",
          "source_id"
        ],
        "title": "AiEdit"
      },
...
      "ContentSource": {
        "type": "string",
        "enum": [
          "manual_edit",
          "ai_edit",
          "initial_generation"
        ],
        "title": "ContentSource"
      },
...
      "GuardrailWarningResponse": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "GuardrailWarningResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "awaiting_start",
          "in_progress",
          "completed",
          "failed"
        ],
        "title": "JobStatus"
      },
...
      "Minute": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Created Datetime"
          },
          "updated_datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Updated Datetime"
          },
          "transcription_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transcription Id"
          },
          "template_name": {
            "type": "string",
            "title": "Template Name",
            "default": "General"
          },
          "user_template_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Template Id"
          },
          "agenda": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agenda"
          }
        },
        "type": "object",
        "required": [
          "transcription_id"
        ],
        "title": "Minute"
      },
      "MinuteListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Created Datetime"
          },
          "updated_datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Updated Datetime"
          },
          "transcription_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transcription Id"
          },
          "template_name": {
            "type": "string",
            "title": "Template Name"
          },
          "agenda": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agenda"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_datetime",
          "updated_datetime",
          "transcription_id",
          "template_name",
          "agenda"
        ],
        "title": "MinuteListItem"
      },
      "MinuteVersion": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Created Datetime"
          },
          "updated_datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Updated Datetime"
          },
          "minute_id": {
            "type": "string",
            "format": "uuid",
            "title": "Minute Id"
          },
          "html_content": {
            "type": "string",
            "title": "Html Content",
            "default": ""
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "awaiting_start"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "template_prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Prompt Version"
          },
          "ai_edit_instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Edit Instructions",
            "description": "If the content source is an AI edit, store the instruction here"
          },
          "content_source": {
            "$ref": "#/components/schemas/ContentSource",
            "default": "initial_generation"
          }
        },
        "type": "object",
        "required": [
          "minute_id"
        ],
        "title": "MinuteVersion"
      },
      "MinuteVersionCreateRequest": {
        "properties": {
          "ai_edit_instructions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiEdit"
              },
              {
                "type": "null"
              }
            ],
            "description": "If the content source is an AI edit, store the instruction and source version id here"
          },
          "content_source": {
            "$ref": "#/components/schemas/ContentSource"
          },
          "html_content": {
            "type": "string",
            "title": "Html Content",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "content_source"
        ],
        "title": "MinuteVersionCreateRequest"
      },
      "MinuteVersionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "minute_id": {
            "type": "string",
            "format": "uuid",
            "title": "Minute Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "created_datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Created Datetime"
          },
          "html_content": {
            "type": "string",
            "title": "Html Content"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "ai_edit_instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Edit Instructions"
          },
          "content_source": {
            "$ref": "#/components/schemas/ContentSource"
          },
          "too_short": {
            "type": "boolean",
            "title": "Too Short",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "minute_id",
          "status",
          "created_datetime",
          "html_content",
          "error",
          "ai_edit_instructions",
          "content_source"
        ],
        "title": "MinuteVersionResponse"
      },
      "MinutesCreateRequest": {
        "properties": {
          "template_name": {
            "type": "string",
            "title": "Template Name",
            "description": "Name of the template to use for the minutes"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id",
            "description": "Optional id of user template"
          },
          "agenda": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Agenda",
            "description": "The agenda for the meeting"
          }
        },
        "type": "object",
        "required": [
          "template_name",
          "template_id"
        ],
        "title": "MinutesCreateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}